+++ indexer/src/actions/dune/daily-contract-usage/client.test.ts
@@ -1,96 +1,112 @@
-//import { DateTime, Duration } from "luxon";
-
 import {
-  resolveDailyContractUsage,
-  IdToAddressMap,
-  DailyContractUsageRawRow,
+  parseDuneContractUsageCSVRow,
+  DuneRawRow,
+  parseDuneCSVArray,
+  SafeAggregate,
+  DailyContractUsageRow,
 } from "./client.js";
+import { parse } from "csv/sync";
+
+// CSV Snippet that's easy to read
+const contractCsvSnippet0 = `
+date,contract_id,usage
+2022-08-28 00:00:00.000 UTC,123,[[0x1111111111111111111111111111111111111111 0x1111111111111111111111111111111111111111 1000000 1] [0x2222222222222222222222222222222222222222 <nil> 2000000000000 2] [<nil> 0x3333333333333333333333333333333333333333 3000000000000000000000000 3]]
+`.trim();
+
+const expectedContractCsvSnippet0: DuneRawRow = {
+  date: "2022-08-28 00:00:00.000 UTC",
+  contract_id: 123,
+  usage: [
+    [
+      "0x1111111111111111111111111111111111111111",
+      "0x1111111111111111111111111111111111111111",
+      "1000000",
+      1,
+    ],
+    ["0x2222222222222222222222222222222222222222", null, "2000000000000", 2],
+    [
+      null,
+      "0x3333333333333333333333333333333333333333",
+      "3000000000000000000000000",
+      3,
+    ],
+  ],
+};
 
 describe("resolveDailyContractUsageResults", () => {
-  const fakeUserIdToAddressMap: IdToAddressMap = {
-    0: "0x0000000000000000000000000000000000000000",
-    1: "0x0000000000000000000000000000000000000001",
-    2: "0x0000000000000000000000000000000000000002",
-  };
-
-  const fakeContractIdToAddressMap: IdToAddressMap = {
-    0: "0xc000000000000000000000000000000000000000",
-    1: "0xc000000000000000000000000000000000000001",
-    2: "0xc000000000000000000000000000000000000002",
-  };
-
-  //const fakeEndDate = DateTime.now();
-  //const fakeStartDate = fakeEndDate.minus(Duration.fromObject({ days: 7 }));
-
-  test("resolve rows correctly", () => {
-    const rows: DailyContractUsageRawRow[] = [
-      {
-        date: "2023-01-01 00:00:00 UTC",
-        contract_id: 0,
-        user_addresses: [],
-        user_ids: [0, 2],
-        contract_total_tx_count: 1,
-        contract_total_l2_gas_cost_gwei: "0",
-        safe_address_count: 0,
-      },
-      {
-        date: "2023-01-01 00:00:00 UTC",
-        contract_id: 1,
-        user_addresses: ["0x0000000000000000000000000000000000000004"],
-        user_ids: [1, 2],
-        contract_total_tx_count: 1,
-        contract_total_l2_gas_cost_gwei: "0",
-        safe_address_count: 0,
-      },
-    ];
-    const resolved = resolveDailyContractUsage(
-      fakeUserIdToAddressMap,
-      fakeContractIdToAddressMap,
-      rows,
-    );
-
-    expect(resolved[0].contractAddress).toBe(
-      "0xc000000000000000000000000000000000000000",
-    );
-    expect(resolved[0].userAddresses.length).toBe(2);
-    expect(resolved[0].userAddresses).toEqual(
-      expect.arrayContaining([
-        "0x0000000000000000000000000000000000000000",
-        "0x0000000000000000000000000000000000000002",
-      ]),
-    );
-
-    expect(resolved[1].contractAddress).toBe(
-      "0xc000000000000000000000000000000000000001",
-    );
-    expect(resolved[1].userAddresses.length).toBe(3);
-    expect(resolved[1].userAddresses).toEqual(
-      expect.arrayContaining([
-        "0x0000000000000000000000000000000000000004",
-        "0x0000000000000000000000000000000000000001",
-        "0x0000000000000000000000000000000000000002",
-      ]),
-    );
+  test("should parse the csv rows from dune", () => {
+    const parsed = parse(contractCsvSnippet0) as string[][];
+    const result = parseDuneContractUsageCSVRow(parsed[1]);
+    expect(result).toEqual(expectedContractCsvSnippet0);
   });
 
-  test("errors when resolving a row with a non-existent user-id", () => {
-    const rows: DailyContractUsageRawRow[] = [
-      {
-        date: "2023-01-01 00:00:00 UTC",
-        contract_id: 0,
-        user_addresses: [],
-        user_ids: [10],
-        contract_total_tx_count: 1,
-        contract_total_l2_gas_cost_gwei: "0",
-        safe_address_count: 0,
-      },
-    ];
+  test("should parse generic csv array from dune", () => {
+    const csvArray0 = "[1234 <nil> test0]";
+    const expectedArray0 = ["1234", "<nil>", "test0"];
+    const csvArray1 = "[4567 test1 <nil>]";
+    const expectedArray1 = ["4567", "test1", "<nil>"];
+    // Nested Array
+    const csvArrayNested0 = `[${csvArray0} ${csvArray1}]`;
+    expect(parseDuneCSVArray(csvArray0)).toEqual(expectedArray0);
+    expect(parseDuneCSVArray(csvArray1)).toEqual(expectedArray1);
+    expect(parseDuneCSVArray(csvArrayNested0)).toEqual([
+      expectedArray0,
+      expectedArray1,
+    ]);
+  });
+
+  test("should throw error when parsing generic csv array from dune that is poorly formatted", () => {
+    const csvArray0 = "[1234 <nil> test0";
+    // Nested Array
     expect(() => {
-      resolveDailyContractUsage(
-        fakeUserIdToAddressMap,
-        fakeContractIdToAddressMap,
-        rows,
-      );
-    }).toThrowError();
+      parseDuneCSVArray(csvArray0);
+    }).toThrow();
+  });
+});
+
+describe("SafeAggregate", () => {
+  const safesToAggregate: DailyContractUsageRow[] = [
+    {
+      date: "2023-10-11T00:00:00Z",
+      contractAddress: "0x1111111111111111111111111111111111111111",
+      userAddress: null,
+      safeAddress: "0x1111111111111111111111111111111111111111",
+      gasCostGwei: "10000",
+      txCount: 11,
+    },
+    {
+      date: "2023-10-11T00:00:00Z",
+      contractAddress: "0x1111111111111111111111111111111111111111",
+      userAddress: null,
+      safeAddress: "0x1111111111111111111111111111111111111111",
+      gasCostGwei: "20000",
+      txCount: 22,
+    },
+    {
+      date: "2023-10-11T00:00:00Z",
+      contractAddress: "0x1111111111111111111111111111111111111111",
+      userAddress: null,
+      safeAddress: "0x1111111111111111111111111111111111111111",
+      gasCostGwei: "30000",
+      txCount: 33,
+    },
+  ];
+
+  test("should aggregrate safe values in the csv", () => {
+    const aggregator = new SafeAggregate(safesToAggregate[0]);
+
+    for (const safe of safesToAggregate.slice(1)) {
+      aggregator.add(safe);
+    }
+    const agg = aggregator.aggregate();
+    expect(agg.contractAddress).toEqual(
+      "0x1111111111111111111111111111111111111111",
+    );
+    expect(agg.safeAddress).toEqual(
+      "0x1111111111111111111111111111111111111111",
+    );
+    expect(agg.userAddress).toBeNull();
+    expect(agg.gasCostGwei).toEqual("60000");
+    expect(agg.txCount).toEqual(66);
   });
 });
